BG-7 - Fixed placeholders rendering underneath pieces
